@@ -2,73 +2,67 @@
   "name": "motoko-playground",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
     "@dfinity/agent": "^2.0.0",
     "@dfinity/candid": "^2.0.0",
     "@dfinity/principal": "^2.0.0",
-    "@monaco-editor/react": "4.4.6",
-    "@types/node": "^14.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/styled-components": "^5.1.10",
+    "@monaco-editor/react": "4.6.0",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "@types/styled-components": "^5.1.34",
     "lodash.debounce": "^4.0.8",
-    "motoko": "^3.6.16",
-    "prettier-plugin-motoko": "^0.8.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "motoko": "^3.8.1",
+    "prettier-plugin-motoko": "^0.9.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-google-charts": "^4.0.1",
-    "react-markdown": "^6.0.2",
-    "react-modal": "^3.14.3",
-    "styled-components": "^5.3.0",
-    "typescript": "^4.3.5",
-    "web-vitals": "^1.0.1"
+    "react-markdown": "^9.0.1",
+    "react-modal": "^3.16.1",
+    "styled-components": "^6.1.13"
   },
   "devDependencies": {
-    "@craco/craco": "^6.0.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.19",
-    "comlink-loader": "^2.0.0",
-    "concurrently": "^7.6.0",
-    "craco-babel-loader": "^1.0.3",
-    "http-proxy-middleware": "^2.0.6",
-    "husky": "^7.0.1",
-    "lint-staged": "^11.1.2",
+    "@svgr/core": "^8.1.0",
+    "@svgr/plugin-jsx": "^8.1.0",
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.14.2",
+    "@vitejs/plugin-react": "^4.3.1",
+    "comlink": "^4.4.1",
+    "http-proxy-middleware": "^3.0.2",
+    "husky": "^9.1.5",
+    "lint-staged": "^15.2.10",
     "mo-dev": "^0.13.0",
-    "prettier": "^2.7.1",
-    "react-scripts": "4.0.3",
-    "wasm-loader": "^1.3.0"
+    "prettier": "^3.3.3",
+    "typescript": "^5.5.4",
+    "vite": "^5.4.3",
+    "dotenv": "^16.3.1",
+    "vite-plugin-environment": "^1.1.3",
+    "vite-plugin-top-level-await": "^1.4.4",
+    "vite-plugin-wasm": "^3.3.0"
   },
   "scripts": {
-    "start": "concurrently -p '|' npm:frontend npm:service",
-    "frontend": "craco start",
+    "start": "vite",
     "service": "mo-dev --deploy",
-    "build": "craco build",
-    "test": "craco test -- --config=config/jest.config.js",
-    "eject": "craco eject",
+    "build": "vite build",
+    "test": "vite test -- --config=config/jest.config.js",
     "deploy": "dfx deploy",
     "clean": "dfx stop && dfx start --clean --background",
     "prestart": "dfx start --background; [ -f .dfx/local/canister_ids.json ] || dfx deploy",
     "predeploy": "npm run prestart",
-    "prepare": "husky install"
+    "prepare": "husky",
+    "dev": "vite"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ],
-    "globals": {
-      "Motoko": true
-    }
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not op_mini all"
-  ],
   "lint-staged": {
     "src/**/*.{js,ts,jsx,tsx,json,css,scss,htm,html,md,mdx}": [
+      "prettier --write"
+    ],
+    "vite.config.ts": [
+      "prettier --write"
+    ],
+    "package.json": [
       "prettier --write"
     ]
   }
