@@ -1,1 +1,12 @@
-/// <reference types="react-scripts" />
+/// <reference types="vite/client" />
+
+declare module "*.svg" {
+  import * as React from "react";
+
+  export const ReactComponent: React.FunctionComponent<
+    React.SVGProps<SVGSVGElement> & { title?: string }
+  >;
+
+  const src: string;
+  export default src;
+}
